% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gather_metrics}
\alias{gather_metrics}
\title{Compute common forecasting performance metrics.}
\usage{
gather_metrics(error_df)
}
\arguments{
\item{error_df}{The data frame returned by ts_cv (see
?ts_cv)}
}
\value{
A data frame with the following columns: model,
horizon, metric, and value. The model column gives the
name of the forecasting model being measured. The
horizon column gives the number of weeks each prediciton
is from the last date in the training data. The metric
column gives the name of the computed performance metric.
It can be any of: rmse, mae, mape, mase, or mamse. The
value column gives the estimated value of the performance
metric.
}
\description{
Computes rmse, mae, mape, mase, and mamse from an
error data frame returned by ts_cv.
}
