% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_key_df}
\alias{get_key_df}
\title{Get the data frame corresponding to a given Primary Key.}
\usage{
get_key_df(multi_key_df, key, add_na = TRUE)
}
\arguments{
\item{multi_key_df}{A data frame of weekly sales data.
Must contain columns for Date and Primary_Key.}

\item{key}{Level of the Primary_Key column to filter for.}

\item{add_na}{Boolean specifying whether to fill missing
weeks with NA.}
}
\value{
A data frame.
}
\description{
Given a multi-key data frame and the desired key,
returns the sub-frame with Primary_Key equal to key.
Can optionally fill missing weeks in the series with NA.
}
\examples{
test_df <- data.frame(Date = c('12/01/2020', '12/07/2020', '12/01/2020'),
                      Primary_Key = c('1.1', '1.1', '1.2'),
                      Weekly_Sales = c(1000, 2000, 1500))
key_df <- get_key_df(test_df, key = '1.2', add_na = FALSE)
key_df_na <- get_key_df(test_df, key = '1.2', add_na = TRUE)
head(key_df)
head(key_df_na)
}
