% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{nest_by_key}
\alias{nest_by_key}
\title{Produce a nested data frame for each key.}
\usage{
nest_by_key(multi_key_df, add_na = TRUE)
}
\arguments{
\item{multi_key_df}{A data frame of weekly sales data.
Must contain columns for Date and Primary_Key.}

\item{add_na}{Boolean specifying whether to fill missing
weeks with NA.}
}
\value{
A data frame two columns: Primary_Key and data.
The data column is a list column that contains a separate
data frame for each level of Primary_Key.
}
\description{
Given a multi-key data frame, nests it by each level
of Primary_Key. Can optionally fill missing weeks in
the series with NA.
}
\examples{
test_df <- data.frame(Date = c('12/01/2020', '12/07/2020', '12/01/2020'),
                      Primary_Key = c('1.1', '1.1', '1.2'),
                      Weekly_Sales = c(1000, 2000, 1500))
nested_df <- nest_by_key(test_df, add_na = FALSE)
nested_df_na <- nest_by_key(test_df, add_na = TRUE)
head(nested_df)
head(nested_df_na)
}
