% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{num_obs}
\alias{num_obs}
\title{Count the number of weeks with sales reported.}
\usage{
num_obs(df)
}
\arguments{
\item{df}{A data frame of weekly sales data. Must
contain a Weekly_Sales column.}
}
\value{
An int giving the number of rows with non-NA
values for the Weekly_Sales column.
}
\description{
Returns the number of rows with non-NA values for the
Weekly_Sales column.
}
\examples{
test_df <- data.frame(Date = c('12/01/2020', '12/07/2020'),
                      Primary_Key = c('1.1', '1.1'),
                      Weekly_Sales = c(1000, NA))
n <- num_obs(test_df)
print(n)
}
