% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess raw weekly sales data.}
\usage{
preprocess(raw_df)
}
\arguments{
\item{raw_df}{A data frame of raw weekly sales data.
Must contain columns named Date, Store, Dept, and
IsHoliday. The Date column must be formated as "\%m/\%d/\%Y".}
}
\value{
A data frame.
}
\description{
Creates new columns for Primary Key, Day, Weekday,
Month, Year, and each holiday (i.e., Christmas,
Thanksgiving, Labor Day, Superbowl and Easter).
Primary Key is the factor interaction between Store
and Department.
}
\examples{
test_df <- data.frame(Date = rep('12/01/2020', times = 10),
                      Store = rep(c('1', '2'), times = 5),
                      Dept = rep(c('3', '4'), times = 5),
                      IsHoliday = rep(0, times = 10))
processed_df <- preprocess(test_df)
head(processed_df)
}
