% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{train_test_split}
\alias{train_test_split}
\title{Split training and test data for modeling.}
\usage{
train_test_split(weekly_df, weeks = 3)
}
\arguments{
\item{weekly_df}{A data frame of weekly sales data.
Must contain a Date column.}

\item{weeks}{Number of weeks to reserve for testing.
Test data will be taken from the last weeks in the
input data frame. Must be less than the total number of
weeks in the input data frame.}
}
\value{
A list of two data frames named train and test.
}
\description{
Split training and test data for modeling.
}
\examples{
test_df <- data.frame(Date = c('12/01/2020', '12/08/2020'),
                      Weekly_Sales = c(1000, 2000))
split_list <- train_test_split(test_df, weeks = 1)
head(split_list$train)
head(split_list$test)
}
