% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ts_cv}
\alias{ts_cv}
\title{Perform time series cross-validation on all models.}
\usage{
ts_cv(data, cutoff_dates, holidays, fit_qgam = FALSE)
}
\arguments{
\item{data}{A data frame of weekly sales data.
Must contain columns for Date, Primary_Key, and
Weekly_Sales. Weeks with missing data should be filled
with NA.}

\item{cutoff_dates}{A list of cutoff dates for creating
train and test splits on each cross-validation loop.}

\item{holidays}{A holiday data frame formatted for use
with Prophet.}
}
\value{
A data frame with the following columns: error,
obs, mean_diff, med_diff, horizon, and model. The error
column is yhat - y for each predicition. It is set to NA
for preditction with no corresponding observation. The
obs column is the corresponding observations. The
mean_diff column is the mean of the diff of the weekly
sales for each set of training data. The med_diff column
is the median instead of the mean. The horizon column
gives the number of weeks each prediciton is from the
last date in the training data. The model column gives
the name of the model that made each prediction.
}
\description{
Given weekly sales data for a single key, and a list of
cutoff dates, trains on data ocurring before the cutoff
dates and tests on data ocurring after. Trains a separate
model for each cutoff date.
}
